+++ Sources/Layout.swift
@@ -2,207 +2,6 @@
 
 import UIKit
 
-@objc public enum FloatingPanelLayoutReferenceGuide: Int {
-    case superview = 0
-    case safeArea = 1
-}
-
-private extension FloatingPanelLayoutReferenceGuide {
-    func layoutGuide(vc: UIViewController) -> LayoutGuideProvider {
-        switch self {
-        case .safeArea:
-            return vc.fp_safeAreaLayoutGuide
-        case .superview:
-            return vc.view
-        }
-    }
-}
-
-@objc public enum FloatingPanelPosition: Int {
-    case top
-    case left
-    case bottom
-    case right
-}
-
-extension FloatingPanelPosition {
-    func mainLocation(_ point: CGPoint) -> CGFloat {
-        switch self {
-        case .top, .bottom: return point.y
-        case .left, .right: return point.x
-        }
-    }
-
-    func mainDimension(_ size: CGSize) -> CGFloat {
-        switch self {
-        case .top, .bottom: return size.height
-        case .left, .right: return size.width
-        }
-    }
-
-    func mainDimensionAnchor(_ layoutGuide: LayoutGuideProvider) -> NSLayoutDimension {
-        switch self {
-        case .top, .bottom: return layoutGuide.heightAnchor
-        case .left, .right: return layoutGuide.widthAnchor
-        }
-    }
-
-    func crossDimension(_ size: CGSize) -> CGFloat {
-        switch self {
-        case .top, .bottom: return size.width
-        case .left, .right: return size.height
-        }
-    }
-
-    func inset(_ insets: UIEdgeInsets) -> CGFloat {
-        switch self {
-        case .top: return insets.top
-        case .left: return insets.left
-        case .bottom: return insets.bottom
-        case .right: return insets.right
-        }
-    }
-}
-
-@objc public enum FloatingPanelReferenceEdge: Int {
-    case top
-    case left
-    case bottom
-    case right
-}
-
-private extension FloatingPanelReferenceEdge {
-    func inset(of insets: UIEdgeInsets) -> CGFloat {
-        switch self {
-        case .top: return insets.top
-        case .left: return insets.left
-        case .bottom: return insets.bottom
-        case .right: return insets.right
-        }
-    }
-    func mainDimension(_ size: CGSize) -> CGFloat {
-        switch self {
-        case .top, .bottom: return size.height
-        case .left, .right: return size.width
-        }
-    }
-}
-
-@objc public protocol FloatingPanelLayoutAnchoring {
-    var referenceGuide: FloatingPanelLayoutReferenceGuide { get }
-    func layoutConstraints(_ fpc: FloatingPanelController, for position: FloatingPanelPosition) -> [NSLayoutConstraint]
-}
-
-@objc final public class FloatingPanelLayoutAnchor: NSObject, FloatingPanelLayoutAnchoring /*, NSCopying */ {
-    @objc public init(absoluteInset: CGFloat, edge: FloatingPanelReferenceEdge, referenceGuide: FloatingPanelLayoutReferenceGuide) {
-        self.inset = absoluteInset
-        self.referenceGuide = referenceGuide
-        self.referenceEdge = edge
-        self.isAbsolute = true
-    }
-
-    @objc public init(fractionalInset: CGFloat, edge: FloatingPanelReferenceEdge, referenceGuide: FloatingPanelLayoutReferenceGuide) {
-        self.inset = fractionalInset
-        self.referenceGuide = referenceGuide
-        self.referenceEdge = edge
-        self.isAbsolute = false
-    }
-    fileprivate let inset: CGFloat
-    fileprivate let isAbsolute: Bool
-    @objc public let referenceGuide: FloatingPanelLayoutReferenceGuide
-    @objc public let referenceEdge: FloatingPanelReferenceEdge
-}
-
-public extension FloatingPanelLayoutAnchor {
-    func layoutConstraints(_ vc: FloatingPanelController, for position: FloatingPanelPosition) -> [NSLayoutConstraint] {
-        let layoutGuide = referenceGuide.layoutGuide(vc: vc)
-        switch position {
-        case .top:
-            return layoutConstraints(layoutGuide, for: vc.surfaceView.bottomAnchor)
-        case .left:
-            return layoutConstraints(layoutGuide, for: vc.surfaceView.rightAnchor)
-        case .bottom:
-            return layoutConstraints(layoutGuide, for:  vc.surfaceView.topAnchor)
-        case .right:
-            return layoutConstraints(layoutGuide, for: vc.surfaceView.leftAnchor)
-        }
-    }
-
-    private func layoutConstraints(_ layoutGuide: LayoutGuideProvider, for edgeAnchor: NSLayoutYAxisAnchor) -> [NSLayoutConstraint] {
-        switch referenceEdge {
-        case .top:
-            if isAbsolute {
-                return [edgeAnchor.constraint(equalTo: layoutGuide.topAnchor, constant: inset)]
-            }
-            let offsetAnchor = layoutGuide.topAnchor.anchorWithOffset(to: edgeAnchor)
-            return [offsetAnchor.constraint(equalTo:layoutGuide.heightAnchor, multiplier: inset)]
-        case .bottom:
-            if isAbsolute {
-                return [layoutGuide.bottomAnchor.constraint(equalTo: edgeAnchor, constant: inset)]
-            }
-            let offsetAnchor = edgeAnchor.anchorWithOffset(to: layoutGuide.bottomAnchor)
-            return [offsetAnchor.constraint(equalTo: layoutGuide.heightAnchor, multiplier: inset)]
-        default:
-            fatalError("Unsupported reference edges")
-        }
-    }
-
-    private func layoutConstraints(_ layoutGuide: LayoutGuideProvider, for edgeAnchor: NSLayoutXAxisAnchor) -> [NSLayoutConstraint] {
-        switch referenceEdge {
-        case .left:
-            if isAbsolute {
-                return [edgeAnchor.constraint(equalTo: layoutGuide.leftAnchor, constant: inset)]
-            }
-            let offsetAnchor = layoutGuide.leftAnchor.anchorWithOffset(to: edgeAnchor)
-            return [offsetAnchor.constraint(equalTo: layoutGuide.widthAnchor, multiplier: inset)]
-        case .right:
-            if isAbsolute {
-                return [layoutGuide.rightAnchor.constraint(equalTo: edgeAnchor, constant: inset)]
-            }
-            let offsetAnchor = edgeAnchor.anchorWithOffset(to: layoutGuide.rightAnchor)
-            return [offsetAnchor.constraint(equalTo: layoutGuide.widthAnchor, multiplier: inset)]
-        default:
-            fatalError("Unsupported reference edges")
-        }
-    }
-}
-
-@objc final public class FloatingPanelIntrinsicLayoutAnchor: NSObject, FloatingPanelLayoutAnchoring /*, NSCopying */ {
-    @objc public init(absoluteOffset offset: CGFloat, referenceGuide: FloatingPanelLayoutReferenceGuide = .safeArea) {
-        self.offset = offset
-        self.referenceGuide = referenceGuide
-        self.isAbsolute = true
-    }
-    // offset = 0.0 -> All content visible
-    // offset = 1.0 -> All content invisible
-    @objc public init(fractionalOffset offset: CGFloat, referenceGuide: FloatingPanelLayoutReferenceGuide = .safeArea) {
-        self.offset = offset
-        self.referenceGuide = referenceGuide
-        self.isAbsolute = false
-    }
-    fileprivate let offset: CGFloat
-    fileprivate let isAbsolute: Bool
-    @objc public let referenceGuide: FloatingPanelLayoutReferenceGuide
-}
-
-public extension FloatingPanelIntrinsicLayoutAnchor {
-    func layoutConstraints(_ vc: FloatingPanelController, for position: FloatingPanelPosition) -> [NSLayoutConstraint] {
-        let surfaceIntrinsicLength = position.mainDimension(vc.surfaceView.intrinsicContentSize)
-        let constant = isAbsolute ? surfaceIntrinsicLength - offset : surfaceIntrinsicLength * (1 - offset)
-        let layoutGuide = referenceGuide.layoutGuide(vc: vc)
-        switch position {
-        case .top:
-            return [vc.surfaceView.bottomAnchor.constraint(equalTo: layoutGuide.topAnchor, constant: constant)]
-        case .left:
-            return [vc.surfaceView.rightAnchor.constraint(equalTo: layoutGuide.leftAnchor, constant: constant)]
-        case .bottom:
-            return [vc.surfaceView.topAnchor.constraint(equalTo: layoutGuide.bottomAnchor, constant: -constant)]
-        case .right:
-            return [vc.surfaceView.leftAnchor.constraint(equalTo: layoutGuide.rightAnchor, constant: -constant)]
-        }
-    }
-}
-
 @objc public protocol FloatingPanelLayout {
     /// The position of the panel in the view of `FloatingPanelController`.
     @objc var position: FloatingPanelPosition { get }
