+++ Sources/LayoutAnchoring.swift
@@ -1,0 +1,117 @@
+// Copyright 2018 the FloatingPanel authors. All rights reserved. MIT license.
+
+import UIKit
+
+@objc public protocol FloatingPanelLayoutAnchoring {
+    var referenceGuide: FloatingPanelLayoutReferenceGuide { get }
+    func layoutConstraints(_ fpc: FloatingPanelController, for position: FloatingPanelPosition) -> [NSLayoutConstraint]
+}
+
+@objc final public class FloatingPanelLayoutAnchor: NSObject, FloatingPanelLayoutAnchoring /*, NSCopying */ {
+    @objc public init(absoluteInset: CGFloat, edge: FloatingPanelReferenceEdge, referenceGuide: FloatingPanelLayoutReferenceGuide) {
+        self.inset = absoluteInset
+        self.referenceGuide = referenceGuide
+        self.referenceEdge = edge
+        self.isAbsolute = true
+    }
+    @objc public init(fractionalInset: CGFloat, edge: FloatingPanelReferenceEdge, referenceGuide: FloatingPanelLayoutReferenceGuide) {
+        self.inset = fractionalInset
+        self.referenceGuide = referenceGuide
+        self.referenceEdge = edge
+        self.isAbsolute = false
+    }
+    let inset: CGFloat
+    let isAbsolute: Bool
+    @objc public let referenceGuide: FloatingPanelLayoutReferenceGuide
+    @objc public let referenceEdge: FloatingPanelReferenceEdge
+}
+
+public extension FloatingPanelLayoutAnchor {
+    func layoutConstraints(_ vc: FloatingPanelController, for position: FloatingPanelPosition) -> [NSLayoutConstraint] {
+        let layoutGuide = referenceGuide.layoutGuide(vc: vc)
+        switch position {
+        case .top:
+            return layoutConstraints(layoutGuide, for: vc.surfaceView.bottomAnchor)
+        case .left:
+            return layoutConstraints(layoutGuide, for: vc.surfaceView.rightAnchor)
+        case .bottom:
+            return layoutConstraints(layoutGuide, for:  vc.surfaceView.topAnchor)
+        case .right:
+            return layoutConstraints(layoutGuide, for: vc.surfaceView.leftAnchor)
+        }
+    }
+
+    private func layoutConstraints(_ layoutGuide: LayoutGuideProvider, for edgeAnchor: NSLayoutYAxisAnchor) -> [NSLayoutConstraint] {
+        switch referenceEdge {
+        case .top:
+            if isAbsolute {
+                return [edgeAnchor.constraint(equalTo: layoutGuide.topAnchor, constant: inset)]
+            }
+            let offsetAnchor = layoutGuide.topAnchor.anchorWithOffset(to: edgeAnchor)
+            return [offsetAnchor.constraint(equalTo:layoutGuide.heightAnchor, multiplier: inset)]
+        case .bottom:
+            if isAbsolute {
+                return [layoutGuide.bottomAnchor.constraint(equalTo: edgeAnchor, constant: inset)]
+            }
+            let offsetAnchor = edgeAnchor.anchorWithOffset(to: layoutGuide.bottomAnchor)
+            return [offsetAnchor.constraint(equalTo: layoutGuide.heightAnchor, multiplier: inset)]
+        default:
+            fatalError("Unsupported reference edges")
+        }
+    }
+
+    private func layoutConstraints(_ layoutGuide: LayoutGuideProvider, for edgeAnchor: NSLayoutXAxisAnchor) -> [NSLayoutConstraint] {
+        switch referenceEdge {
+        case .left:
+            if isAbsolute {
+                return [edgeAnchor.constraint(equalTo: layoutGuide.leftAnchor, constant: inset)]
+            }
+            let offsetAnchor = layoutGuide.leftAnchor.anchorWithOffset(to: edgeAnchor)
+            return [offsetAnchor.constraint(equalTo: layoutGuide.widthAnchor, multiplier: inset)]
+        case .right:
+            if isAbsolute {
+                return [layoutGuide.rightAnchor.constraint(equalTo: edgeAnchor, constant: inset)]
+            }
+            let offsetAnchor = edgeAnchor.anchorWithOffset(to: layoutGuide.rightAnchor)
+            return [offsetAnchor.constraint(equalTo: layoutGuide.widthAnchor, multiplier: inset)]
+        default:
+            fatalError("Unsupported reference edges")
+        }
+    }
+}
+
+@objc final public class FloatingPanelIntrinsicLayoutAnchor: NSObject, FloatingPanelLayoutAnchoring /*, NSCopying */ {
+    @objc public init(absoluteOffset offset: CGFloat, referenceGuide: FloatingPanelLayoutReferenceGuide = .safeArea) {
+        self.offset = offset
+        self.referenceGuide = referenceGuide
+        self.isAbsolute = true
+    }
+    // offset = 0.0 -> All content visible
+    // offset = 1.0 -> All content invisible
+    @objc public init(fractionalOffset offset: CGFloat, referenceGuide: FloatingPanelLayoutReferenceGuide = .safeArea) {
+        self.offset = offset
+        self.referenceGuide = referenceGuide
+        self.isAbsolute = false
+    }
+    let offset: CGFloat
+    let isAbsolute: Bool
+    @objc public let referenceGuide: FloatingPanelLayoutReferenceGuide
+}
+
+public extension FloatingPanelIntrinsicLayoutAnchor {
+    func layoutConstraints(_ vc: FloatingPanelController, for position: FloatingPanelPosition) -> [NSLayoutConstraint] {
+        let surfaceIntrinsicLength = position.mainDimension(vc.surfaceView.intrinsicContentSize)
+        let constant = isAbsolute ? surfaceIntrinsicLength - offset : surfaceIntrinsicLength * (1 - offset)
+        let layoutGuide = referenceGuide.layoutGuide(vc: vc)
+        switch position {
+        case .top:
+            return [vc.surfaceView.bottomAnchor.constraint(equalTo: layoutGuide.topAnchor, constant: constant)]
+        case .left:
+            return [vc.surfaceView.rightAnchor.constraint(equalTo: layoutGuide.leftAnchor, constant: constant)]
+        case .bottom:
+            return [vc.surfaceView.topAnchor.constraint(equalTo: layoutGuide.bottomAnchor, constant: -constant)]
+        case .right:
+            return [vc.surfaceView.leftAnchor.constraint(equalTo: layoutGuide.rightAnchor, constant: -constant)]
+        }
+    }
+}
